removing unnecessary Digests singleton; null pattern on digests
This has proven to be not needed.  Not really adding any value over
just setting a class var on the main Dassets module.

I also moved to using the null pattern on the digests file.  The
digests will return a NullDigestsFile obj until `init` is called.
On `init`, the digests file configured will be read into memory.