have an ENV var drive what file the CLI loads to configure Dassets
It defaults to `config/assets`.  This gets the Dassets CLI more in-line
with how the Sanford CLI works.  The CLI expects that whatever file
is specified by `ENV['DASSETS_CONFIG_FILE']` configures Dassets
appropriately.

Moving to this design simplified a lot of things and revealed a
whole bunch of "cruft" in the CLI that just wasn't needed.  Good
times.