updating the assets file handling to follow the Sanford pattern
Define a config for the assets file, default it with an ENV var.  Then
require in the assets file on `init`.

This updates Dassets to this pattern - the goal is to be as similar
to Sanford as possible.  The reason for this is they are both
gems, with CLIs, that need some configs in an expected file, and then
do stuff.