redesign: read source list on init
This adds source list handling.  On init, the source list is read
and filtered based on the config.  By default, only files are
included in the source path list and any paths in the output path
are removed.  In addition the source filter is applied to the path
list.

This is the new way Dassets knows which files to digest or host
with the middleware.