cleanups to digests modeling
Removed some unneeded methods and tried to simplify the api.  This
is prepping for the new compile behavior.  The digests stores the
previous compile state and is needed for pruning old compiled files
from the output on compile.

This is, of course, in addition to its asset file handling for the
middleware.