Switch to using `config.output_path` instead of deprecated `files_path`
This rename and reuse more aptly describes the nature of the files
in play.  These are the files that have been output by the digest
process.

In addition, this renames the asset file `cache_path` property to
`url`.  Cache path really doesn't describe the intent of this property
from the asset file's perspective.  From the asset file's perspective
this property is the URL used to compose the files `href`.  What
you do with the property (ie use it to cache or whatever) is not the
asset file's concern.