find a source file by its digest path
This adds a utility class method for finding source files by their
digest paths.  This is useful because it will allow asset files
(which already know their digest path) to now know their source
files.

Source files are found by scanning the sources and finding the
last source with a matching digest path.  We choose the last one
to handle the case where their are two source files that produce
the same digest path.  When digested, the last source file would
overwrite any previous sources.  Therefore the last source file
matching that digest path is the intended one.

This also introduces the concept of a `NullSourceFile` to handle
the case where a matching source file cannot be found.