rename asset file `md5` to `fingerprint`
No need to name the attr after the algorithm used to create it.  A
more appropriate name is "fingerprint" - it doesn't matter how you
compute that fingerprint.

This will allow us to more easily alter the method we use compute
file fingerprints and is a more appropriate name.

Also included here: I removed the `AssetFile.from_abs_path` class
method.  It is no longer being used and needed to be cleaned up.