source files now handle output paths
Source files will only write (ie cache) digested output if an output
path is configured.  No output path is configured by default.  All
output path handling is removed from the asset file.

An example of using this would be to only configure you "public" dir
as the  output path in non-development modes.  With this you would always
be serving freshly digested source in development. But in non-development
you would digest the source on the first request and then serve
out of the public dir directly every request after that.