have the middleware response digest the asset file on success
This will write compiled output to the file store on success responses.

If you make your file store the public directory, you will essentially
"populate" it on the fly.  The first request hits the middelware
and write the file to the public dir.  Any subsequent requests just
get the file served from the public dir directly.