more forgiving Dassets initialization
Don't raise load errors on requiring the assets file - just quietly
move on as this may not be necessary in all scenarios.  Just raise
a more specific runtime error when missing required configs.

The goal here is to allow programmatic configuration of Dassets not
using the asset file.  This may be desired in certain init scenarios.