combinations: rename source cache to source proxy
This is a more accurate name for the purpose of this object.  It
proxies the source file(s) from the asset file that accesses their
data.  It abstract and handles the source files themselves (with
soon-to-be-combination handling), source data access, and cacheing.

No behavior changes - although this did expose some law-of-demeter
violations that I cleaned up both in the request and dassets_tests.