combinations: allow registering combinations at config time
This adds a combination registration api to the config.  You add
combinations by giving the combination digest path and the list of
source digest paths it maps to.

This will be used by the proxy to 'look up' the source digest paths
for the given digest path.  Combinations return the keyed digest path
in the case no combination has been registered.  This allows digest
paths to 'pass thru' if not registered as a combination.