cleanups to the existing test suite
This adds no new behavior.  These are cleanups to the test suite
to get it up to modern Assert testing style/standards.

In the process of cleaning up, I noticed the file store tests were
not very robust.  I updated the tests and in the process changed
the internal API for accessing a "null store" object.