auto init Dassets
Dassets requires an init step to set the asset files collection to
an empty hash.  This is to avoid a conditional in the index operator
method (`[]`).  Previously, the developer would have to remember to
call this method before using the index operator and if they forgot
Dassets would throw a nasty nil error.

This switches to just auto initializing Dassets on require.  There
is really no advantage to forcing a dev to do this and it just
results in confusion and errors if they don't.  This removes all
that aggrivation.