updated README - fill in remaining TODO sections
This has been needed for a while now as the API and behavior has
become fairly stable.  The goal here is to explain a few of the
more complex/abstract concepts in Dassets.

Romo is key here as it uses all of these concepts and provides a
tidy example of them in a production type setting.