source proxy: cleanup tests; add tests for empty combos
We discovered an edge-case type bug in how `mtime` was being calculated
on source proxies.  If you source proxy was for a combo that contained
an empty combo, `mtime` would error.  This is b/c the `mtime` of an
empty combo in `nil` so when your combo that contains the empty combo
tries to calculate its `mtime`, it ends up comparing `nil` to the
`mtime`s of its other source files which errors.

This fixes the issue by ignoring the `mtime` of emtpy combinations.
We do this by just compacting the `mtime` values array before selecting
the max value.

This also basically rewrites the source proxy tests.  There are a
number of small changes there:

* update tests to match implementation order
* break our separate contexts for the different scenarios being tested
* add contexts for specifically testing empty combos and combos composing
  empty combos (covers the error scenario above)
* update test descriptions to be more specific about what they are
  testing