add config `set_base_url` method; use to cleanup test state
This fixes an issue where modified base url state was being set
in the request tests but not being properly reset.  This issue is
that the `base_url` method won't change the base url if given a nil
value.  The requests tests were resetting the base url to the orig
value but the orig value was nil so the reset had no effect.

This updates the config api to have a `set_base_url` method that
sets *any* given value.  The `base_url` method uses this method
to set its values and the request test now use this method to
force reset the base url, ensuring clean state.

This was noticed when I went to update the asset file's `href`
method to honor any configured base url.