update AssetFile#url to honor any configured `base_url`
Since the base url will be honored (and needed) when hosting asset
files, don't make the user manually add it to asset urls.  Instead
have the base url be auto-included when generating asset file urls.

Note: in addition, I've renamed the old `url` method to `path` as
this name more accurately describes the intent of the method: that
this is the path relative to the asset root for this asset.  This
value is also used to cache the assets contents to the file store
(which is confusing if the method is named `url`).  The new `url`
method is more accurately a "url", honoring the base url as needed.
I've aliased `href` to `url` for backwards compatibility.