Update asset file to write to the file store using its url
This updates the asset file to write to the file store using its
url. The goal here is to make sure the base url is used when
writing to the file store. Previously, the asset files were being
written without the base url. This caused HTTP servers to not be
able to directly serve the files because the url (which included
the base url) didn't match the path where the files were stored.

Now that the path method is no longer being used, its being merged
into the url method. This is to avoid any future confusion over
using the url method with the file store and not the path method.
Related to this, the readme has also been updated to better explain
how the file store should be configured. Its intended to be
configured with the public dir so HTTP servers can serve files out
of it.

This also updates the test suite to add additional testing using
the base url. The digest system and unit tests now will randomly
have a base url. This ensure digesting works with or without a
base url being configured. Because of this, the digested file1
support file can't be checked in. Without a base url the tests
would always generate the same digested file1 but now that a base
url may be used they won't. The tests now remove the files they
generate which means they will also remove the checked in version.

This also fixes a require issue with the server request and a
Gemfile typo. Requiring `rack/request` for some reason didn't
define the `Rack::Request::GET` constant and the server request
unit tests would fail.