Allow configuring a content and fingerprint cache separately
This changes the cache options to allow configuring a separate
content and fingerprint cache. This allows configuring no-cache
for the content (so file contents aren't stored in memory) but
using a mem-cache for fingerprints. This is useful when using
Dassets in a production environment when the assets are cached to
a public directory. The assets content doesn't need to be cached
because an HTTP server can serve them directly but the fingerprints
are cached so asset urls can still be generated quickly.

The config now has a separate option for content caching and
fingerprint caching. These are automatically passed to an asset
files source proxy just like the previous single cache was. The
fingerprint cache is only used to store fingerprints and the
content cache is only used to store file contents.