remove ns-options as a dependency
This is part of getting this gem working in modern ruby versions.
We are moving away from using ns-options as a dependency b/c it
has performance issues when accessing options and on top of that
its API hasn't proven to be especially useful over a struct-like
class.

This removes the one option defined (`file_store`) using
ns-options logic.  Now, this option is handled manually but
behaves almost identically. There is one backwards incompatible
change though: there is no longer a `file_store=` writer.  I
updated the test suite to account for this.  I also checked our
apps and we seem to only be using the DSL method to set the file
store in them so this shouldn't break things.