rework how Dassets stores its source files for faster lookups
This switches to storing the source files in a Hash rather than
an array.  The goal is to make `SourceFile.find_by_digest_path`
lookups have linear performance as the number of source files
increases.  Using a hash removes the need to scan all of the files
in the sources looking for a match.

This also switches to building the SourceFile objects stored by
the source files hash on `init`.  This means that the source files
are now "compiled" and prepped on init so that any future lookups
has linear performance.

This is all to make Dassets better (and more performant) at
fronting/hosting large sets of source files.