(hotfix) update the FileStore with better var names and tests
This changes all came about b/c I used the FileStore as reference
for another FileStore implementation in another gem.  This updates
the `url` var name to be `url_path` b/c it is expected that this
var just has the path portion of a url.  This also updates the
FileStore tests to do similar var name changes but to also get
them up to our latest conventions.