error loudly if requesting a digest path that can't be digested
The goal here is to help find errors in Dassets configuration and
setup more easily.  Before, if you messed up your Dassets config
or if you typo'd a digest path, Dassets would just return an
"empty" asset file as kind of a null pattern placeholder for the
asset being requested.  The developer wouldn't realize there was
a problem until a request came for the asset and returned a 404.
This 404 would cause the page to render incorrectly but give no
clue to the dev that there was a problem with Dassets.  The only
clue the dev would have is the asset path would end with a dash
and no digest fingerprint.

This isn't very helpful.  A better solution would be for Dassets
to error loudly with detailed error info so the dev would clearly
know the problem is with Dassets and have some info to help debug
the problem.  This error can now be caught in automated test
suites or reported via production exception trackers, etc.

So, this updates the Dassets index operator `[]` to now check the
asset file it will return for a fingerprint.  If it has no
fingerprint that means there was a problem digesting the asset
file.  Either there is no source file for that digest path or
there is no combination for that digest path or there is a combo
but the combo has an incorrect source file definition, etc.

In this case, Dassets raises a custom exception with the digest
path requested.  In addition, it outputs the defined combinations
with sorted combo digest paths and also outputs the sorted digest
paths of the configured source files.  This info can be used to
debug missing/typo'd digest paths (which is most likely the reason
for the problem.