add `asset_file` method that doesn't error to compliment `[]` method
We need a method that looks up the asset file from a digest path
so that the request class can use it.  Requests should never hard
error in this situation b/c they need to pass on the request if
the request is not for an asset file.  If it errors, you never get
a chance to check if the request was for an asset file or not.

This switches the request to use `asset_file` method that doesn't
error and reverts the request tests to their pre-PR-83 logic.
Changing the request tests in PR 83 should have been a red-flag
that the request behavior was changing but I blew right past that
warning.  Oh well, this cleans things up.

Note: this also fixes a bug where the hard error msg from PR 83
won't build if you have a combination with no values.  I noticed
this while debugging this problem.